move out run_algo into ann_benchmarks.runner
